pyproject.toml
[project]
name = "arxiv-opportunity-finder"
version = "0.1.0"
description = "A tool for finding research paper implementation opportunities"
authors = [
    {name = "Gajesh",email = "gajesh@eigenlabs.org"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arxiv (>=2.1.3,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "flask (>=3.1.0,<4.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "feedparser (>=6.0.11,<7.0.0)",
    "sgmllib3k (>=1.0.0,<2.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "openai (>=1.65.2,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "tenacity (>=9.0.0,<10.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pypdf (>=5.3.1,<6.0.0)",
    "anthropic (>=0.49.0,<0.50.0)",
    "aiohttp (>=3.11.13,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pdfplumber (>=0.10.4,<0.11.0)",
    "flask-cors (>=5.0.1,<6.0.0) ; python_version >= \"3.10\" and python_version < \"4.0\""
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
